Adds Locale "en" for groups in archives
MMMM in the SimpleDateFormat constructor is replaced with the month name
of the default locale. The date:longDate filter of Selmer in Archives.html
generates a format with english month names. On non-english Systems that
leads to months names in the local language for the group names, but
english month names for the full dates on the post. To be consistent, the
locale "en" gets set manually on the SimpleDateFormatter constructor call.
This results in english month names on the archives page.