server.clj: enable auto-refresh in the fast mode
Enable auto-reloading of the blog's changed page in the browser if in the fast serve mode.

It only makes sense in the fast mode because the ring refresh middleware kicks in soon after it detects any change in the output dir - likely a while before Cryogen gets to compiling the page the user is currently viewing/editing - and thus it will try to load a url that it not yet available, resulting in Not Found.